+++ example/lib/data_stream.dart
@@ -25,7 +25,7 @@ class DataStreamExample extends StatelessWidget {
         child: CustomScrollView(
           slivers: [
             SliverStreamBuilder<String>(
-              sliverBuilder: (_, delegate) => SliverGrid(
+              sliverBuilder: (_, delegate, {key}) => SliverGrid(
                 delegate: delegate,
                 gridDelegate: const SliverGridDelegateWithMaxCrossAxisExtent(
                   maxCrossAxisExtent: 100,