+++ lib/localization/strings.g.dart
@@ -1,0 +1,359 @@
+
+/*
+ * Generated file. Do not edit.
+ *
+ * Locales: 2
+ * Strings: 4 (2.0 per locale)
+ *
+ * Built on 2022-02-07 at 21:20 UTC
+ */
+
+import 'package:flutter/widgets.dart';
+
+const AppLocale _baseLocale = AppLocale.en;
+AppLocale _currLocale = _baseLocale;
+
+/// Supported locales, see extension methods below.
+///
+/// Usage:
+/// - LocaleSettings.setLocale(AppLocale.en) // set locale
+/// - Locale locale = AppLocale.en.flutterLocale // get flutter locale from enum
+/// - if (LocaleSettings.currentLocale == AppLocale.en) // locale check
+enum AppLocale {
+	en, // 'en' (base locale, fallback)
+	ru, // 'ru'
+}
+
+/// Method A: Simple
+///
+/// No rebuild after locale change.
+/// Translation happens during initialization of the widget (call of L).
+///
+/// Usage:
+/// String a = L.someKey.anotherKey;
+/// String b = L['someKey.anotherKey']; // Only for edge cases!
+StringsEn _L = _currLocale.translations;
+StringsEn get L => _L;
+
+/// Method B: Advanced
+///
+/// All widgets using this method will trigger a rebuild when locale changes.
+/// Use this if you have e.g. a settings page where the user can select the locale during runtime.
+///
+/// Step 1:
+/// wrap your App with
+/// TranslationProvider(
+/// 	child: MyApp()
+/// );
+///
+/// Step 2:
+/// final L = Translations.of(context); // Get L variable.
+/// String a = L.someKey.anotherKey; // Use L variable.
+/// String b = L['someKey.anotherKey']; // Only for edge cases!
+class Translations {
+	Translations._(); // no constructor
+
+	static StringsEn of(BuildContext context) {
+		final inheritedWidget = context.dependOnInheritedWidgetOfExactType<_InheritedLocaleData>();
+		if (inheritedWidget == null) {
+			throw 'Please wrap your app with "TranslationProvider".';
+		}
+		return inheritedWidget.translations;
+	}
+}
+
+class LocaleSettings {
+	LocaleSettings._(); // no constructor
+
+	/// Uses locale of the device, fallbacks to base locale.
+	/// Returns the locale which has been set.
+	static AppLocale useDeviceLocale() {
+		final locale = AppLocaleUtils.findDeviceLocale();
+		return setLocale(locale);
+	}
+
+	/// Sets locale
+	/// Returns the locale which has been set.
+	static AppLocale setLocale(AppLocale locale) {
+		_currLocale = locale;
+		_L = _currLocale.translations;
+
+		if (WidgetsBinding.instance != null) {
+			// force rebuild if TranslationProvider is used
+			_translationProviderKey.currentState?.setLocale(_currLocale);
+		}
+
+		return _currLocale;
+	}
+
+	/// Sets locale using string tag (e.g. en_US, de-DE, fr)
+	/// Fallbacks to base locale.
+	/// Returns the locale which has been set.
+	static AppLocale setLocaleRaw(String rawLocale) {
+		final locale = AppLocaleUtils.parse(rawLocale);
+		return setLocale(locale);
+	}
+
+	/// Gets current locale.
+	static AppLocale get currentLocale {
+		return _currLocale;
+	}
+
+	/// Gets base locale.
+	static AppLocale get baseLocale {
+		return _baseLocale;
+	}
+
+	/// Gets supported locales in string format.
+	static List<String> get supportedLocalesRaw {
+		return AppLocale.values
+			.map((locale) => locale.languageTag)
+			.toList();
+	}
+
+	/// Gets supported locales (as Locale objects) with base locale sorted first.
+	static List<Locale> get supportedLocales {
+		return AppLocale.values
+			.map((locale) => locale.flutterLocale)
+			.toList();
+	}
+}
+
+/// Provides utility functions without any side effects.
+class AppLocaleUtils {
+	AppLocaleUtils._(); // no constructor
+
+	/// Returns the locale of the device as the enum type.
+	/// Fallbacks to base locale.
+	static AppLocale findDeviceLocale() {
+		final String? deviceLocale = WidgetsBinding.instance?.window.locale.toLanguageTag();
+		if (deviceLocale != null) {
+			final typedLocale = _selectLocale(deviceLocale);
+			if (typedLocale != null) {
+				return typedLocale;
+			}
+		}
+		return _baseLocale;
+	}
+
+	/// Returns the enum type of the raw locale.
+	/// Fallbacks to base locale.
+	static AppLocale parse(String rawLocale) {
+		return _selectLocale(rawLocale) ?? _baseLocale;
+	}
+}
+
+// context enums
+
+// interfaces generated as mixins
+
+// translation instances
+
+late StringsEn _translationsEn = StringsEn.build();
+late StringsRu _translationsRu = StringsRu.build();
+
+// extensions for AppLocale
+
+extension AppLocaleExtensions on AppLocale {
+
+	/// Gets the translation instance managed by this library.
+	/// [TranslationProvider] is using this instance.
+	/// The plural resolvers are set via [LocaleSettings].
+	StringsEn get translations {
+		switch (this) {
+			case AppLocale.en: return _translationsEn;
+			case AppLocale.ru: return _translationsRu;
+		}
+	}
+
+	/// Gets a new translation instance.
+	/// [LocaleSettings] has no effect here.
+	/// Suitable for dependency injection and unit tests.
+	///
+	/// Usage:
+	/// final t = AppLocale.en.build(); // build
+	/// String a = t.my.path; // access
+	StringsEn build() {
+		switch (this) {
+			case AppLocale.en: return StringsEn.build();
+			case AppLocale.ru: return StringsRu.build();
+		}
+	}
+
+	String get languageTag {
+		switch (this) {
+			case AppLocale.en: return 'en';
+			case AppLocale.ru: return 'ru';
+		}
+	}
+
+	Locale get flutterLocale {
+		switch (this) {
+			case AppLocale.en: return const Locale.fromSubtags(languageCode: 'en');
+			case AppLocale.ru: return const Locale.fromSubtags(languageCode: 'ru');
+		}
+	}
+}
+
+extension StringAppLocaleExtensions on String {
+	AppLocale? toAppLocale() {
+		switch (this) {
+			case 'en': return AppLocale.en;
+			case 'ru': return AppLocale.ru;
+			default: return null;
+		}
+	}
+}
+
+// wrappers
+
+GlobalKey<_TranslationProviderState> _translationProviderKey = GlobalKey<_TranslationProviderState>();
+
+class TranslationProvider extends StatefulWidget {
+	TranslationProvider({required this.child}) : super(key: _translationProviderKey);
+
+	final Widget child;
+
+	@override
+	_TranslationProviderState createState() => _TranslationProviderState();
+
+	static _InheritedLocaleData of(BuildContext context) {
+		final inheritedWidget = context.dependOnInheritedWidgetOfExactType<_InheritedLocaleData>();
+		if (inheritedWidget == null) {
+			throw 'Please wrap your app with "TranslationProvider".';
+		}
+		return inheritedWidget;
+	}
+}
+
+class _TranslationProviderState extends State<TranslationProvider> {
+	AppLocale locale = _currLocale;
+
+	void setLocale(AppLocale newLocale) {
+		setState(() {
+			locale = newLocale;
+		});
+	}
+
+	@override
+	Widget build(BuildContext context) {
+		return _InheritedLocaleData(
+			locale: locale,
+			child: widget.child,
+		);
+	}
+}
+
+class _InheritedLocaleData extends InheritedWidget {
+	final AppLocale locale;
+	Locale get flutterLocale => locale.flutterLocale; // shortcut
+	final StringsEn translations; // store translations to avoid switch call
+
+	_InheritedLocaleData({required this.locale, required Widget child})
+		: translations = locale.translations, super(child: child);
+
+	@override
+	bool updateShouldNotify(_InheritedLocaleData oldWidget) {
+		return oldWidget.locale != locale;
+	}
+}
+
+// pluralization feature not used
+
+// helpers
+
+final _localeRegex = RegExp(r'^([a-z]{2,8})?([_-]([A-Za-z]{4}))?([_-]?([A-Z]{2}|[0-9]{3}))?$');
+AppLocale? _selectLocale(String localeRaw) {
+	final match = _localeRegex.firstMatch(localeRaw);
+	AppLocale? selected;
+	if (match != null) {
+		final language = match.group(1);
+		final country = match.group(5);
+
+		// match exactly
+		selected = AppLocale.values
+			.cast<AppLocale?>()
+			.firstWhere((supported) => supported?.languageTag == localeRaw.replaceAll('_', '-'), orElse: () => null);
+
+		if (selected == null && language != null) {
+			// match language
+			selected = AppLocale.values
+				.cast<AppLocale?>()
+				.firstWhere((supported) => supported?.languageTag.startsWith(language) == true, orElse: () => null);
+		}
+
+		if (selected == null && country != null) {
+			// match country
+			selected = AppLocale.values
+				.cast<AppLocale?>()
+				.firstWhere((supported) => supported?.languageTag.contains(country) == true, orElse: () => null);
+		}
+	}
+	return selected;
+}
+
+// translations
+
+// Path: <root>
+class StringsEn {
+
+	/// You can call this constructor and build your own translation instance of this locale.
+	/// Constructing via the enum [AppLocale.build] is preferred.
+	StringsEn.build();
+
+	/// Access flat map
+	dynamic operator[](String key) => _flatMap[key];
+
+	// Internal flat map initialized lazily
+	late final Map<String, dynamic> _flatMap = _buildFlatMap();
+
+	// ignore: unused_field
+	late final StringsEn _root = this;
+
+	// Translations
+	String get genericLoadingError => 'The error has been acquired.';
+	String get retryButton => 'Retry';
+}
+
+// Path: <root>
+class StringsRu extends StringsEn {
+
+	/// You can call this constructor and build your own translation instance of this locale.
+	/// Constructing via the enum [AppLocale.build] is preferred.
+	StringsRu.build()
+		: super.build();
+
+	/// Access flat map
+	@override dynamic operator[](String key) => _flatMap[key];
+
+	// Internal flat map initialized lazily
+	late final Map<String, dynamic> _flatMap = _buildFlatMap();
+
+	// ignore: unused_field
+	@override late final StringsRu _root = this;
+
+	// Translations
+	@override String get genericLoadingError => 'Произошла ошибка при загрузке данных.';
+	@override String get retryButton => 'Попробовать снова';
+}
+
+/// Flat map(s) containing all translations.
+/// Only for edge cases! For simple maps, use the map function of this library.
+
+extension on StringsEn {
+	Map<String, dynamic> _buildFlatMap() {
+		return {
+			'genericLoadingError': 'The error has been acquired.',
+			'retryButton': 'Retry',
+		};
+	}
+}
+
+extension on StringsRu {
+	Map<String, dynamic> _buildFlatMap() {
+		return {
+			'genericLoadingError': 'Произошла ошибка при загрузке данных.',
+			'retryButton': 'Попробовать снова',
+		};
+	}
+}